LayoutTests/http/tests/from-origin/fetch-iframe-from-origin-same-blocked-expected.txt

CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/from-origin/resources/fetch.php?fromOrigin=same due to access control checks.
CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8000/from-origin/resources/fetch.php?fromOrigin=same due to access control checks.
Tests that a cross-origin fetch in an iframe load fails if the server blocks cross-origin loads with a 'From-Origin: same' response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Localhost fetch in 127.0.0.1 iframe blocked and 127.0.0.1 fetch in localhost iframe blocked.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
The iframe making a localhost fetch call.

--------
Frame: '<!--frame2-->'
--------
The iframe making a 127.0.0.1 fetch call.
